refactor(products): drop browse-card registration pill
The inline status pill on browse cards is gone — registration state is now
carried by the CTA button (View / disabled Full / hidden) and, for muni
clubs, the seat-availability bar. Splits the surviving useRegistrationCta
hook out of the deleted registration-pill.tsx into registration-cta.ts.

Removes the now-orphaned pill i18n keys across all locales, the pill demo
section in the UI components page, and updates the products-architecture
doc. Demo schedule fixtures corrected to drop the (EET) abbrev — a Helsinki
viewer of Helsinki-authored products sees no tz decoration.
