+++ messages/en.json
@@ -1324,16 +1324,9 @@
       "notAvailableInCurrency": "Not in {currency}",
       "seatsCapacity": "{count, plural, one {# seat} other {# seats}}",
       "waitlistAvailable": "Waitlist available",
-      "endedBadge": "Ended",
       "endedNote": "This one's wrapped — keep an eye out for the next one.",
       "viewDetails": "View",
       "manage": "Manage",
-      "pillSpotsLeft": "{count, plural, one {Only # spot left} other {Only # spots left}}",
-      "pillNeedsMore": "{count, plural, one {Need # more to start} other {Need # more to start}}",
-      "pillFull": "Full",
-      "pillFullWaitlist": "Full — waitlist open",
-      "pillOpensOn": "Opens {date}",
-      "pillStarted": "Already started",
       "fullDisabled": "Full",
       "online": "Online"
     },
