+++ messages/fi.json
@@ -1324,16 +1324,9 @@
       "notAvailableInCurrency": "Ei valuutassa {currency}",
       "seatsCapacity": "{count, plural, one {# paikka} other {# paikkaa}}",
       "waitlistAvailable": "Jonotuslista käytössä",
-      "endedBadge": "Päättynyt",
       "endedNote": "Tämä on päättynyt – seuraava on tulossa pian.",
       "viewDetails": "Katso",
       "manage": "Hallitse",
-      "pillSpotsLeft": "{count, plural, one {Vain # paikka jäljellä} other {Vain # paikkaa jäljellä}}",
-      "pillNeedsMore": "{count, plural, one {# ilmoittautuminen puuttuu} other {# ilmoittautumista puuttuu}}",
-      "pillFull": "Täynnä",
-      "pillFullWaitlist": "Täynnä – jonotuslista auki",
-      "pillOpensOn": "Aukeaa {date}",
-      "pillStarted": "Alkanut",
       "fullDisabled": "Täynnä",
       "online": "Verkossa"
     },
