+++ messages/sv.json
@@ -1324,16 +1324,9 @@
       "notAvailableInCurrency": "Inte i {currency}",
       "seatsCapacity": "{count, plural, one {# plats} other {# platser}}",
       "waitlistAvailable": "Väntelista finns",
-      "endedBadge": "Avslutad",
       "endedNote": "Den här är slut — håll utkik efter nästa.",
       "viewDetails": "Visa",
       "manage": "Hantera",
-      "pillSpotsLeft": "{count, plural, one {Bara # plats kvar} other {Bara # platser kvar}}",
-      "pillNeedsMore": "{count, plural, one {# till behövs för att starta} other {# till behövs för att starta}}",
-      "pillFull": "Fullt",
-      "pillFullWaitlist": "Fullt — väntelista öppen",
-      "pillOpensOn": "Öppnar {date}",
-      "pillStarted": "Redan igång",
       "fullDisabled": "Fullt",
       "online": "Online"
     },
