+++ messages/tlh.json
@@ -1369,16 +1369,9 @@
       "notAvailableInCurrency": "{currency} ghajbe'",
       "seatsCapacity": "{count, plural, one {# quS} other {# quS}}",
       "waitlistAvailable": "loS tetlh tu'lu'",
-      "endedBadge": "rIn",
       "endedNote": "rInpu' – veb yIloS.",
       "viewDetails": "yIlegh",
       "manage": "yIche'",
-      "pillSpotsLeft": "{count, plural, one {neH # quS Sumlu'} other {neH # quS Sumlu'}}",
-      "pillNeedsMore": "{count, plural, one {tagh, # latlh poQ} other {tagh, # latlh poQ}}",
-      "pillFull": "buy'",
-      "pillFullWaitlist": "buy' — loS tetlh",
-      "pillOpensOn": "{date} poSmoH",
-      "pillStarted": "taghta'",
       "fullDisabled": "buy'",
       "online": "QumDaq"
     },
