+++ src/app/(dashboard)/admin/ui-components/page.tsx
@@ -61,9 +61,7 @@ import {
   ProductBrowseCardView,
   type ProductBrowseCardViewProps,
 } from "@/components/public/products/product-browse-card-view";
-import { RegistrationPill } from "@/components/public/products/registration-pill";
 import { SeatAvailabilityBar } from "@/components/public/products/seat-availability-bar";
-import type { RegistrationState } from "@/components/public/products/derive-registration-state";
 import { SignupPanel } from "@/components/public/products/signup-panel";
 import {
   buildDetailFixture,
@@ -959,42 +957,8 @@ function buildLoadedSessions(
 }
 
 /* ------------------------------------------------------------------ */
-/*  Products (browse + purchased cards, registration pill)             */
+/*  Products (browse + purchased cards)                                 */
 /* ------------------------------------------------------------------ */
-
-// Every state where the pill earns a row. Default-open is intentionally
-// excluded — the pill returns null in that case (the Sign-up button
-// alone says everything a parent needs).
-const DEMO_REGISTRATION_STATES: { label: string; state: RegistrationState }[] = [
-  {
-    label: "Almost full · 2 spots left",
-    state: { kind: "open", seatCount: 8, seatsLeft: 2, waitlistEnabled: false },
-  },
-  {
-    label: "Needs more sign-ups (threshold)",
-    state: { kind: "pending_thr", threshold: 6, count: 2 },
-  },
-  {
-    label: "Full · waitlist open",
-    state: { kind: "full_waitlist", seatCount: 8 },
-  },
-  {
-    label: "Full · closed",
-    state: { kind: "full_closed", seatCount: 8 },
-  },
-  {
-    label: "Sign-ups open later",
-    state: { kind: "closed_pre", opensAt: "2026-05-15T00:00:00Z" },
-  },
-  {
-    label: "Already started (camp/event)",
-    state: { kind: "running_late" },
-  },
-  {
-    label: "Ended",
-    state: { kind: "ended" },
-  },
-];
 
 // One sample of each price shape so the price block is exercised across
 // the demo browse cards.
@@ -1011,14 +975,14 @@ const SAMPLE_PRICE_FREE: ProductBrowseCardViewProps["price"] = { kind: "free" };
 // Browse-card display props for each state we want to render in full.
 const BROWSE_DEMO_CARDS: { label: string; props: ProductBrowseCardViewProps }[] = [
   {
-    label: "Default · plenty of seats (no pill)",
+    label: "Default · plenty of seats",
     props: {
       name: "Tuesday Minecraft Builders",
       description:
         "Weekly creative sessions where your gamer collaborates with new friends and a Gedu who really gets it.",
       imagePath: null,
       topicLabel: "MINECRAFT",
-      scheduleLines: ["Tuesday · 17:00–18:30 (EET)"],
+      scheduleLines: ["Tuesday · 17:00–18:30"],
       ageLine: "Ages 8–12",
       seatsHint: { kind: "capacity", count: 8 },
       locationLine: { kind: "online", label: "Online" },
@@ -1035,7 +999,7 @@ const BROWSE_DEMO_CARDS: { label: string; props: ProductBrowseCardViewProps }[] 
         "Build, race, and collab with your crew — small group with regular faces every week.",
       imagePath: null,
       topicLabel: "ROBLOX",
-      scheduleLines: ["Wednesday · 17:00–18:30 (EET)"],
+      scheduleLines: ["Wednesday · 17:00–18:30"],
       ageLine: "Ages 9–13",
       seatsHint: { kind: "capacity", count: 8 },
       locationLine: { kind: "in_person", label: "Tapiolan koulu" },
@@ -1052,7 +1016,7 @@ const BROWSE_DEMO_CARDS: { label: string; props: ProductBrowseCardViewProps }[] 
         "Group challenge that runs once enough builders sign up — gather your crew and we'll lock in a start date.",
       imagePath: null,
       topicLabel: "ROBLOX",
-      scheduleLines: ["24–28 March (EET)", "10:00–14:00"],
+      scheduleLines: ["24–28 March", "10:00–14:00"],
       ageLine: "Ages 9–14",
       seatsHint: null,
       locationLine: { kind: "online_muni", label: "Espoo" },
@@ -1069,7 +1033,7 @@ const BROWSE_DEMO_CARDS: { label: string; props: ProductBrowseCardViewProps }[] 
         "Drop-in event for parents and gamers — light competition, lots of laughter.",
       imagePath: null,
       topicLabel: "FORTNITE",
-      scheduleLines: ["Friday 12 April · 18:00–20:00 (EET)"],
+      scheduleLines: ["Friday 12 April · 18:00–20:00"],
       ageLine: "Ages 10+",
       seatsHint: { kind: "capacity", count: 12 },
       locationLine: { kind: "in_person", label: "Iso Omena" },
@@ -1086,7 +1050,7 @@ const BROWSE_DEMO_CARDS: { label: string; props: ProductBrowseCardViewProps }[] 
         "A week-long story-driven adventure across multiple games. Sign-ups open soon.",
       imagePath: null,
       topicLabel: "ADVENTURE",
-      scheduleLines: ["12–16 August (EET)", "10:00–15:00"],
+      scheduleLines: ["12–16 August", "10:00–15:00"],
       ageLine: "Ages 9–13",
       seatsHint: { kind: "capacity", count: 16 },
       locationLine: { kind: "in_person", label: "Sogverse HQ" },
@@ -1103,7 +1067,7 @@ const BROWSE_DEMO_CARDS: { label: string; props: ProductBrowseCardViewProps }[] 
         "Already underway — late joins aren't supported once a camp is running.",
       imagePath: null,
       topicLabel: "ROBLOX",
-      scheduleLines: ["20–24 April (EET)", "10:00–14:00"],
+      scheduleLines: ["20–24 April", "10:00–14:00"],
       ageLine: "Ages 8–12",
       seatsHint: { kind: "capacity", count: 10 },
       locationLine: { kind: "in_person", label: "Ressun peruskoulu" },
@@ -1119,7 +1083,7 @@ const BROWSE_DEMO_CARDS: { label: string; props: ProductBrowseCardViewProps }[] 
       description: "This event has wrapped — keep an eye out for the next one.",
       imagePath: null,
       topicLabel: "FORTNITE",
-      scheduleLines: ["Saturday 22 March · 14:00–17:00 (EET)"],
+      scheduleLines: ["Saturday 22 March · 14:00–17:00"],
       ageLine: "Ages 10+",
       seatsHint: null,
       locationLine: { kind: "in_person", label: "Tampere-talo" },
@@ -1138,7 +1102,7 @@ const BROWSE_DEMO_CARDS: { label: string; props: ProductBrowseCardViewProps }[] 
         "Weekly in-school sessions funded by the municipality — free for families to register.",
       imagePath: null,
       topicLabel: "Minecraft",
-      scheduleLines: ["Thursday · 15:00–16:30 (EET)"],
+      scheduleLines: ["Thursday · 15:00–16:30"],
       ageLine: "Ages 10–12",
       seatsHint: { kind: "capacity", count: 15 },
       locationLine: { kind: "online", label: "Online" },
@@ -1224,37 +1188,13 @@ function SeatAvailabilityDemo() {
 
 function ProductsDemo() {
   return (
-    <div className="space-y-8">
-      <SubSection title="Pill — every state">
-        <p className="text-sm text-muted-foreground mb-3">
-          The pill speaks parent voice and only renders when there&rsquo;s something
-          actionable or urgency-creating to say. Default-open (&ldquo;you can sign up&rdquo;)
-          gets no pill — the Sign-up button alone says everything a parent needs.
-        </p>
-        <div className="flex flex-wrap items-center gap-x-6 gap-y-3">
-          {DEMO_REGISTRATION_STATES.map(({ label, state }) => (
-            <div key={label} className="flex flex-col gap-1.5">
-              <DemoCaption>{label}</DemoCaption>
-              <RegistrationPill state={state} />
-            </div>
-          ))}
+    <div className="grid gap-x-6 gap-y-8 sm:grid-cols-2 lg:grid-cols-3">
+      {BROWSE_DEMO_CARDS.map(({ label, props }) => (
+        <div key={label} className="flex flex-col gap-2">
+          <DemoCaption>{label}</DemoCaption>
+          <ProductBrowseCardView {...props} />
         </div>
-      </SubSection>
-
-      <SubSection title="Browse cards (in context)">
-        <p className="text-sm text-muted-foreground mb-4">
-          Full card with the pill inline next to the topic label. Each example
-          renders one of the registration states the deriver returns.
-        </p>
-        <div className="grid gap-x-6 gap-y-8 sm:grid-cols-2 lg:grid-cols-3">
-          {BROWSE_DEMO_CARDS.map(({ label, props }) => (
-            <div key={label} className="flex flex-col gap-2">
-              <DemoCaption>{label}</DemoCaption>
-              <ProductBrowseCardView {...props} />
-            </div>
-          ))}
-        </div>
-      </SubSection>
+      ))}
     </div>
   );
 }
@@ -1802,9 +1742,10 @@ export default function AdminUIComponentsPage() {
       <Section title="Products — Browse Cards">
         <p className="text-sm text-muted-foreground -mt-2">
           Parent-facing card surface for products in the shop (/shop). Every
-          product renders the same browse card whether or not the customer
-          owns it. The registration pill speaks parent voice and only appears
-          when there&rsquo;s something actionable to say.
+          product renders the same browse card whether or not the customer owns
+          it. One card per registration state the deriver returns — the CTA
+          button and (for muni clubs) the seat-availability bar carry the state
+          signal.
         </p>
         <ProductsDemo />
       </Section>
